requirements.txt
asgiref==3.3.4
Django==3.2.3
django-cors-headers==3.7.0
django-filter==2.4.0
django-polymorphic==3.0.0
djangorestframework==3.12.4
drf-generators==0.5.0
Markdown==3.3.4
pytz==2021.1
sqlparse==0.4.1
